requirements.txt
# Core Dependencies
boto3~=1.34.0
botocore~=1.34.0

# YAML Processing
pyyaml>=6.0.1

# Async Support
aiohttp>=3.9.0

# Monitoring and Logging
python-json-logger>=2.0.7
structlog>=23.2.0

# CLI Dependencies
rich>=13.7.0
click>=8.1.7

# Development Dependencies
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.1
ruff>=0.1.6
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pre-commit>=3.5.0

# Type Stubs
types-PyYAML>=6.0.12.12
boto3-stubs[essential]>=1.34.0
types-setuptools>=69.0.0
